✨ Add Slack client
Client handles formatting the restaurant info and posting it to a Slack
API. The endpoint used is determined by the SLACK_URL environment
variable set in the .env file.

There is a dummy version for development that returns a response object
without actually posting anything to an external API.

This is a small application, so tying the slack module to the restaurant
info is justified -- no need to build separate formatter and API client.