🐛 Use response content instead of text
Content is properly encoded to unicode characters, so asserting works
when expected string contains non-ASCII characters.

Bug was found when correcting a test bug to not expect a menu for a
specific day -- even with a fixture, this only works on the weekday in
question since the client correctly finds the menu for the current day.

Also update the integration test to check that at least the restaurant
name was found in parsing the real content.