🐳 Add Dockerfile for packaging the app
Add building the container to the build script.

The container does not include tests but the build script runs the test
suite before building the app.

Environment variables are not packaged in the container; give the env
file as a run-time argument, e.g.:

docker run --rm --env-file=.qa.env lunch-slack